extracted AbstractPlcConnection class
